Move torchvision models to submodule
