+++ segmentation_robustness_framework/models/__init__.py
@@ -1,8 +1,10 @@
 # Base segmentation models
 from .base_model import SegmentationModel
 
-# Segmentation models
-from .deeplab import DeepLabV3
-from .fcn import FCN
+# Torchvision segmentation models
+from .torchvision_models.fcn import TorchvisionFCN
+from .torchvision_models.deeplabv3 import TorchvisionDeepLabV3
 
-__all__ = ["SegmentationModel", "DeepLabV3", "FCN"]
+# Other models below ...
+
+__all__ = ["SegmentationModel", "TorchvisionDeepLabV3", "TorchvisionFCN"]
