+++ segmentation_robustness_framework/models/torchvision_models/__init__.py
@@ -1,0 +1,2 @@
+from .fcn import TorchvisionFCN
+from .deeplabv3 import TorchvisionDeepLabV3
